Use `request-method' as the default value for `request-vars'' source
Using `both' as the default value for `request-vars'' `source' keyword
params is dangerous.  Consider the following example (in awful):

  (use awful)

  (enable-sxml #t)

  (define-page (main-page-path)
    (lambda ()
      (with-request-variables (user admin)
        (if user
            `(ul
              (li ,user)
              (li ,(or admin "--")))
            `(form (@ (method "post"))
                   (input (@ (type "text") (name "user")))
                   (input (@ (type "submit")))))))
    method: '(get post))

An attacker could maliciously make an user follow a link to the form
with the query string set to "?admin=bar".  If `source' is bound to
`both', `request-vars' will read from both the query string and the
request body, leading to user=<what user filled> and admin=bar.

This change sets the default value for `source' to `request-method',
that is, `request-vars' will read the query string or the request body
depending on the request method (never both).

With this change, in the example mentioned above, if an attacker makes
an user follow a link to the form with the query string set to
"?admin=bar", `request-vars' will read th request body _only_, since
the form method is "post".  Thus, the handler will get user=<what user
filled> and bar=#f.

Thanks to Peter Bex for the heads-up and discussions on this issue.